Make rtps writer impl private (#4869)
* Refs #20821: Make StatefulPersistentWriter private


* Refs #20821: Make StatelessPersistentWriter private


* Refs #20821: Make StatefulWriter private


* Refs #20821: Make StatelessWriter private


* Refs #20821: Apply suggestions: Make persistenteWriter private


* Refs #20821. Avoid min / max macros on ReaderProxyTests.


---------
