+++ src/cpp/rtps/builtin/discovery/endpoint/EDPServer.cpp
@@ -23,8 +23,9 @@
 #include <fastdds/rtps/attributes/HistoryAttributes.h>
 #include <fastdds/rtps/attributes/ReaderAttributes.h>
 #include <fastdds/rtps/history/WriterHistory.h>
-#include <fastdds/rtps/writer/StatefulWriter.h>
+
 #include <rtps/reader/StatefulReader.hpp>
+#include <rtps/writer/StatefulWriter.hpp>
 
 using namespace ::eprosima::fastrtps::rtps;
