+++ src/cpp/rtps/builtin/discovery/endpoint/EDPSimple.cpp
@@ -33,7 +33,6 @@
 #include <fastdds/rtps/builtin/data/WriterProxyData.h>
 #include <fastdds/rtps/history/ReaderHistory.h>
 #include <fastdds/rtps/history/WriterHistory.h>
-#include <fastdds/rtps/writer/StatefulWriter.h>
 
 #include <rtps/builtin/BuiltinProtocols.h>
 #include <rtps/builtin/discovery/endpoint/EDPSimpleListeners.h>
@@ -42,6 +41,7 @@
 #include <rtps/history/TopicPayloadPoolRegistry.hpp>
 #include <rtps/history/TopicPayloadPoolRegistry.hpp>
 #include <rtps/reader/StatefulReader.hpp>
+#include <rtps/writer/StatefulWriter.hpp>
 #ifdef FASTDDS_STATISTICS
 #include <statistics/rtps/monitor-service/interfaces/IProxyObserver.hpp>
 #endif //FASTDDS_STATISTICS
