+++ src/cpp/rtps/participant/RTPSParticipantImpl.cpp
@@ -41,10 +41,6 @@
 #include <fastdds/rtps/transport/TCPv4TransportDescriptor.h>
 #include <fastdds/rtps/transport/TCPv6TransportDescriptor.h>
 #include <fastdds/rtps/transport/UDPv4TransportDescriptor.h>
-#include <fastdds/rtps/writer/StatefulPersistentWriter.h>
-#include <fastdds/rtps/writer/StatefulWriter.h>
-#include <fastdds/rtps/writer/StatelessPersistentWriter.h>
-#include <fastdds/rtps/writer/StatelessWriter.h>
 #include <fastdds/utils/IPFinder.h>
 
 #include <rtps/builtin/BuiltinProtocols.h>
@@ -64,6 +60,10 @@
 #include <rtps/reader/StatefulReader.hpp>
 #include <rtps/reader/StatelessPersistentReader.hpp>
 #include <rtps/reader/StatelessReader.hpp>
+#include <rtps/writer/StatefulWriter.hpp>
+#include <rtps/writer/StatefulPersistentWriter.hpp>
+#include <rtps/writer/StatelessPersistentWriter.hpp>
+#include <rtps/writer/StatelessWriter.hpp>
 #include <statistics/rtps/GuidUtils.hpp>
 #include <utils/Semaphore.hpp>
 #include <utils/string_utilities.hpp>
