+++ src/cpp/statistics/rtps/monitor-service/MonitorService.hpp
@@ -30,11 +30,11 @@
 #include <fastdds/dds/publisher/qos/WriterQos.hpp>
 #include <fastdds/rtps/common/Guid.h>
 #include <fastdds/rtps/history/WriterHistory.h>
-#include <fastdds/rtps/writer/StatefulWriter.h>
 
 #include "Interfaces.hpp"
 #include <rtps/history/ITopicPayloadPool.h>
 #include <rtps/resources/TimedEvent.h>
+#include <rtps/writer/StatefulWriter.hpp>
 #include <statistics/rtps/monitor-service/MonitorServiceListener.hpp>
 #include <statistics/types/monitorservice_types.hpp>
 #include <statistics/types/monitorservice_typesPubSubTypes.h>
