+++ test/unittest/dds/publisher/DataWriterTests.cpp
@@ -36,9 +36,6 @@
 #include <fastdds/dds/subscriber/qos/DataReaderQos.hpp>
 #include <fastdds/dds/subscriber/qos/SubscriberQos.hpp>
 #include <fastdds/dds/subscriber/Subscriber.hpp>
-#include <fastdds/rtps/writer/RTPSWriter.h>
-#include <fastdds/rtps/writer/StatefulWriter.h>
-
 #include <fastdds/publisher/DataWriterImpl.hpp>
 
 #include "../../common/CustomPayloadPool.hpp"
